Change tab navigator into expo tab navigator, and fix party page
